+++ modules/webgl/src/adapter/converters/device-parameters.ts
@@ -151,6 +151,18 @@ export function setDeviceParameters(device: Device, parameters: Parameters) {
   //   // Handled by depthBias
   // },
 
+  if (parameters.provokingVertex) {
+    if (device.features.has('provoking-vertex-webgl')) {
+      const extensions = webglDevice.getExtension('EXT_provoking_vertex');
+      const ext = extensions['EXT_provoking_vertex'];
+      const vertex = map('provokingVertex', parameters.provokingVertex, {
+        first: GL.FIRST_VERTEX_CONVENTION,
+        last: GL.LAST_VERTEX_CONVENTION
+      });
+      ext?.provokeVertex(vertex);
+    }
+  }
+
   // DEPTH STENCIL
 
   if (parameters.depthWriteEnabled !== undefined) {